refactor(cap-service): decreased test files into 1 file
