--- test/cap-service/config/prettier.test.ts
@@ -1,74 +1,0 @@
-/**
- * Test files test the whole results of the Projen project creation
- * configuration orchestration. E.g. the npm test module does test the result of the whole `package.json` file
- * and that the devcontainer config module scripts are created. The devcontainer test module itself will not test
- * if the related devcontainer npm scripts are created.
- *
- * The `setup.ts` file includes all relevant bootstrap steps for the test files.
- * It is important that each test file imports and uses the `snapshot` otherwise the bootstrap will not run for this test file.
- **/
-
-import * as prettier from '../../shared/prettier';
-import { snapshot } from './setup';
-
-/**
- * We use not a describe block here because the jest test runner `Test Results` pane will show the test names.
- * The `Testing` pane shows a hierarchy but for this hierarchy we already have the test file names available.
- */
-test('Prettier settings are set properly', (): void => {
-  const expectedSettings = {
-    '//': '~~ Generated by projen. To modify, edit .projenrc.ts and run "npx projen".',
-    overrides: [
-      {
-        files: '*.*',
-        options: {
-          semi: true,
-          trailingComma: 'all',
-          singleQuote: true,
-          printWidth: 120,
-          tabWidth: 2,
-        },
-      },
-    ],
-  };
-  prettier.testSettings(snapshot, expectedSettings);
-});
-
-test('Ignore patterns matches expected content', (): void => {
-  const expectedEntries: string[] = [
-    '/.commitlintrc.ts',
-    '/.devcontainer.json',
-    '/.gitattributes',
-    '/.github/ISSUE_TEMPLATE/bug.yml',
-    '/.github/ISSUE_TEMPLATE/feature.yml',
-    '/.github/ISSUE_TEMPLATE/housekeeping.yml',
-    '/.github/ISSUE_TEMPLATE/question.yml',
-    '/.github/ISSUE_TEMPLATE/story.yml',
-    '/.github/pull_request_template.md',
-    '/.github/workflows/deployment.yml',
-    '/.github/workflows/release.yml',
-    '/.gitignore',
-    '/.husky/commit-msg',
-    '/.husky/pre-commit',
-    '/.husky/pre-push',
-    '/.prettierignore',
-    '/.prettierrc.json',
-    '/.vscode/settings.json',
-    '/.vscode/launch.json',
-    '/.vscode/tasks.json',
-    '/@cds-models',
-    '/API.md',
-    '/CHANGELOG.md',
-    '/coverage/',
-    '/cliff.toml',
-    '/dist',
-    '/eslint.config.mjs',
-    '/gen',
-    '/junit.xml',
-    '/package-lock.json',
-    '/package.json',
-    '/tsconfig.json',
-    '/test-reports/',
-  ];
-  prettier.testIgnore(snapshot, expectedEntries);
-});
