--- test/cap-service/config/vscode.test.ts
@@ -1,37 +1,0 @@
-/**
- * Test files test the whole results of the Projen project creation
- * configuration orchestration. E.g. the npm test module does test the result of the whole `package.json` file
- * and that the devcontainer config module scripts are created. The devcontainer test module itself will not test
- * if the related devcontainer npm scripts are created.
- *
- * The `setup.ts` file includes all relevant bootstrap steps for the test files.
- * It is important that each test file imports and uses the `snapshot` otherwise the bootstrap will not run for this test file.
- **/
-
-import * as vscode from '../../shared/vscode';
-import { snapshot } from './setup';
-
-/**
- * We use not a describe block here because the jest test runner `Test Results` pane will show the test names.
- * The `Testing` pane shows a hierarchy but for this hierarchy we already have the test file names available.
- */
-test('VsCode settings are set properly', (): void => {
-  const expectedSettings: Record<string, string | boolean | number> = {
-    '//': '~~ Generated by projen. To modify, edit .projenrc.ts and run "npx projen".',
-    'editor.tabSize': 2,
-    'editor.stickyTabStops': true,
-    'typescript.inlayHints.parameterNames.enabled': 'all',
-    'typescript.inlayHints.enumMemberValues.enabled': true,
-    'typescript.inlayHints.variableTypes.enabled': true,
-    'typescript.inlayHints.propertyDeclarationTypes.enabled': true,
-    'javascript.inlayHints.parameterNames.suppressWhenArgumentMatchesName': false,
-    'javascript.inlayHints.variableTypes.suppressWhenTypeMatchesName': false,
-    'typescript.inlayHints.functionLikeReturnTypes.enabled': true,
-    'typescript.inlayHints.parameterTypes.enabled': true,
-    'editor.inlayHints.fontSize': 10,
-    'editor.inlayHints.padding': true,
-    'editor.formatOnSave': true,
-    'editor.formatOnPaste': true,
-  };
-  vscode.testSettings(snapshot, expectedSettings);
-});
